changed price from uint to float

+++ models/book.go
@@ -1,8 +1,8 @@
 package models
 
 type Book struct {
-	ID     string `param:"id" json:"id"`
-	Title  string `json:"title"`
-	Author string `json:"author"`
-	Price  uint   `json:"price"`
+	ID     string  `param:"id" json:"id"`
+	Title  string  `json:"title"`
+	Author string  `json:"author"`
+	Price  float64 `json:"price"`
 }
